requirements-test-withpsycopg.txt
# test runs requirements (versions we'll be testing against) - automatically updated by requires.io
psycopg2-binary==2.9.1; platform_python_implementation != "PyPy"
psycopg2cffi==2.9.0; platform_python_implementation == "PyPy"
-r requirements-test.txt
